modal moved to portal
Modal in portal for correct positioning.
Added transition on buttons hover.
File management according to our team style.
Added name to exit button
