+++ src/components/LogoutModal/LogoutModal.jsx
@@ -1,12 +1,21 @@
-import { useSelector } from 'react-redux';
 import ButtonsSet from '../Button/Button';
-import { ModalContainer, Icon, Title, ExitBtn } from './LogoutModal.styled';
+import ReactDOM from 'react-dom';
+import { Route } from 'react-router-dom';
+import { ModalContainer, Icon, Title, ExitBtn } from './styled/LogoutModal.styled';
 import icons from '../../images/icons.svg';
 
-const LogoutModal = ({ onAction }) => {
-  // const isLoggedIn =useSelector(({auth})=>auth.isLoggedIn);
-  const isLoggedIn = true;
+//Создания портала и вызов модалки для правильного позиционирования
+const Modal = ({onAction,value}) => {
+  return ReactDOM.createPortal(
+    <LogoutModal onAction={onAction} value={value} />,
+    document.getElementById('portal')
+  )
+}
 
+//Модалка с пропсами
+const LogoutModal = ({ onAction,value }) => {
+  const logout = value === "logout"
+ 
   return (
     <ModalContainer>
       <ExitBtn onClick={e => onAction()}>
@@ -14,7 +23,7 @@ const LogoutModal = ({ onAction }) => {
           <use href={icons + '#icon-close'}></use>
         </Icon>
       </ExitBtn>
-      {isLoggedIn ? (
+      {logout ? (
         <Title>Вы действительно хотите выйти?</Title>
       ) : (
         <Title>Вы уверены?</Title>
@@ -23,4 +32,4 @@ const LogoutModal = ({ onAction }) => {
     </ModalContainer>
   );
 };
-export default LogoutModal;
+export default Modal;
